--- a/browser/plugins/console-breadcrumbs.js
+++ b/browser/plugins/console-breadcrumbs.js
@@ -1,4 +1,4 @@
-const { map, filter } = require('../../base/lib/es-utils')
+const { map, reduce, filter } = require('../../base/lib/es-utils')
 
 /*
  * Leaves breadcrumbs when console log methods are called
@@ -8,19 +8,19 @@ module.exports = {
     map(CONSOLE_LOG_METHODS, method => {
       const original = console[method]
       console[method] = (...args) => {
-        client.leaveBreadcrumb('Console output', {
-          severity: /^group/.test(method) ? 'log' : method,
-          message: map(args, arg => {
-            // do the best/simplest stringification of each argument
-            let stringified = String(arg)
-            // unless it stringifies to [object Object], use the toString() value
-            if (stringified !== '[object Object]') return stringified
-            // otherwise attempt to JSON stringify (with indents/spaces)
-            try { stringified = JSON.stringify(arg, null, 2) } catch (e) {}
-            // any errors, fallback to [object Object]
-            return stringified
-          }).join('\n')
-        }, 'log')
+        client.leaveBreadcrumb('Console output', reduce(args, (accum, arg, i) => {
+          // do the best/simplest stringification of each argument
+          let stringified = String(arg)
+          // if it stringifies to [object Object] attempt to JSON stringify
+          if (stringified === '[object Object]') {
+            // catch stringify errors and fallback to [object Object]
+            try { stringified = JSON.stringify(arg) } catch (e) {}
+          }
+          accum[`[${i}]`] = stringified
+          return accum
+        }, {
+          severity: method.indexOf('group') === 0 ? 'log' : method
+        }), 'log')
         original.apply(console, args)
       }
       console[method]._restore = () => { console[method] = original }