… dist/bugsnag.js